Revert "outside click bug fixed"
